RealmImplementation finish. Change id Expense to long and correct changes in SummaryYear List

+++ app/src/main/java/org/upv/ccupeiro/contadroid/actualmonth/view/presenter/ActualMonthPresenter.java
@@ -50,7 +50,7 @@ public class ActualMonthPresenter {
             view.showSaveError();
         }
     }
-    public void deleteExpense(int id){
+    public void deleteExpense(long id){
         if(deleteExpense.execute(id)){
             view.redrawTabs();
             view.showDeleteCorrect();
@@ -59,7 +59,7 @@ public class ActualMonthPresenter {
         }
     }
 
-    public void changePaidStatus(int id, boolean paid){
+    public void changePaidStatus(long id, boolean paid){
         if(changePaidStatus.execute(id,paid)){
             view.redrawTabs();
         }else{
